requirements.txt
asgiref==3.3.1
Django==2.2.10
django-extensions==3.1.0
djangorestframework==3.12.2
pytz==2020.5
sqlparse==0.4.1



